requirements.txt
twilio==8.12.0
groq==0.4.2
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
websockets==12.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
elevenlabs==0.2.27
deepgram-sdk==2.12.0
livekit-agents[deepgram,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-multipart==0.0.15
httpx==0.24.1
soundfile==0.12.1
numpy==1.26.4
scipy==1.12.0